app/enterprise_data_search/requirements.txt
Flask
flask-cors
gunicorn
google-cloud-datastore
google-cloud-storage
google-generativeai
google-cloud-vision